+++ Dependencies/MelonStartScreen/UI/Objects/UI_Object.cs
@@ -1,0 +1,7 @@
+﻿namespace MelonLoader.MelonStartScreen.UI.Objects
+{
+    internal abstract class UI_Object
+    {
+        internal abstract void Render();
+    }
+}