+++ Dependencies/MelonStartScreen/UI/UIAnchor.cs
@@ -1,0 +1,20 @@
+﻿namespace MelonLoader.MelonStartScreen.UI
+{
+    internal enum UIAnchor
+    {
+        // Upper
+        UpperLeft,
+        UpperCenter,
+        UpperRight,
+
+        // Middle
+        MiddleLeft,
+        MiddleCenter,
+        MiddleRight,
+
+        // Lower
+        LowerLeft,
+        LowerCenter,
+        LowerRight
+    }
+}